Adding default 120 secs timout to pebble.ProcessPool when running exonerate()
